+++ TraZer/Tracer/include/tzFlatMeshTriangle.h
@@ -22,7 +22,7 @@ class tzFlatMeshTriangle: public tzMeshTriangle {
 		operator= (const tzFlatMeshTriangle& rhs);
 				
 		virtual	bool
-		hit(const tzRay& ray, double& tmin, tzShadeRec& sr) const; 																	
+		hit(const tzRay& ray, float& tmin, tzShadeRec& sr) const;
 
 };
