+++ TraZer/Tracer/include/tzSmoothMeshTriangle.h
@@ -22,7 +22,7 @@ class tzSmoothMeshTriangle: public tzMeshTriangle {
 		operator= (const tzSmoothMeshTriangle& rhs);
 				
 		virtual	bool 
-		hit(const tzRay& ray, double& tmin, tzShadeRec& sr) const; 																	
+		hit(const tzRay& ray, float& tmin, tzShadeRec& sr) const;
 
 	protected:
