+++ TraZer/Tracer/include/tzSmoothTriangle.h
@@ -27,9 +27,9 @@ class tzSmoothTriangle: public tzIGeometricObject {
 		get_bounding_box(void);
 		
 		virtual bool 
-		hit(const tzRay& ray, double& tmin, tzShadeRec& sr) const;		 					 
+		hit(const tzRay& ray, float& tmin, tzShadeRec& sr) const;
 		
-		virtual bool shadowHit(const tzRay& ray, double& tmin) const;
+		virtual bool shadowHit(const tzRay& ray, float& tmin) const;
 		
 	private:
