+++ TraZer/Tracer/include/tzTriangle.h
@@ -29,9 +29,9 @@ class tzTriangle: public tzIGeometricObject {
 		compute_normal(void);
 		
 		virtual bool 
-		hit(const tzRay& ray, double& tmin, tzShadeRec& sr) const;		 					 
+		hit(const tzRay& ray, float& tmin, tzShadeRec& sr) const;
 		
-		virtual bool shadowHit(const tzRay& ray, double& tmin) const;
+		virtual bool shadowHit(const tzRay& ray, float& tmin) const;
 		
 	private:
